requirements.txt
streamlit>=1.35.0
supabase>=2.4.0
google-generativeai>=0.7.0
reducto>=0.1.0
python-dotenv>=1.0.0
pandas>=2.0.0
plotly>=5.18.0
